requirements.txt
google-adk>=1.2.0
anthropic>=0.40.0
boto3>=1.35.0
python-dotenv>=1.0.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit